Updated README.md following suggestions
Alex Georgopoulos suggested on the help-cfengine mailing list that
/etc/systemd/system should be preferred to the /usr/local/lib/systemd/system
directory, since there is no trace of it in the man page systemd.unit
upstream.
https://groups.google.com/d/msg/help-cfengine/WnXQqUl8xJI/xpNLoV9yAwAJ

That is true:
http://www.freedesktop.org/software/systemd/man/systemd.unit.html

though it's mentioned in the man page of systemd itself:
http://www.freedesktop.org/software/systemd/man/systemd.html

Both directories are somehow "officially recognized". However, the docs
converge on /etc/systemd/system so I decided to change the README
accordingly and to add a footnote about the "local" directory.